image-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pillow>=9.0.0
opencv-python>=4.8.0
pillow-avif-plugin>=1.3.0
python-multipart>=0.0.6
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
